fix error generating weekly reports

+++ src/Models/Apis/Anomalies.hs
@@ -354,7 +354,7 @@ SELECT id, created_at, updated_at, project_id, acknowleged_at, anomaly_type, tar
       WHEN anomaly_type='shape' THEN (select (count(*), COALESCE(max(created_at), iss.created_at)) from apis.request_dumps where project_id=iss.project_id AND shape_hash=iss.target_hash AND created_at > current_timestamp - interval '14d' )
       WHEN anomaly_type='format' THEN (select (count(*), COALESCE(max(created_at), iss.created_at)) from apis.request_dumps where project_id=iss.project_id AND iss.target_hash=ANY(format_hashes) AND created_at > current_timestamp - interval '14d' )
       WHEN anomaly_type='runtime_exception' THEN (select (count(*), COALESCE(max(created_at), iss.created_at)) from apis.request_dumps where project_id=iss.project_id AND errors @> ('[{"hash": "' || iss.target_hash || '"}]')::jsonb AND created_at > current_timestamp - interval '14d')
-      ELSE (0, NOW()::TEXT)
+      ELSE (0::BIGINT, NOW()::TEXT)
     END as req_count
     FROM apis.issues iss WHERE project_id = ? and target_hash=? LIMIT 1
     |]
@@ -374,7 +374,7 @@ getReportAnomalies pid report_type = query Select (Query $ encodeUtf8 q) pid
       -- Format requires a CONTAINS query which is not covered by the regular indexes. GIN index can't have created_at compound indexes, so its a slow query
       -- WHEN anomaly_type='format' THEN (select (count(*), COALESCE(max(created_at), iss.created_at)) from apis.request_dumps where project_id=iss.project_id AND iss.target_hash=ANY(format_hashes) AND created_at > current_timestamp - interval '14d' )
       WHEN anomaly_type='runtime_exception' THEN (select (count(*), COALESCE(max(created_at), iss.created_at)) from apis.request_dumps where project_id=iss.project_id AND errors @> ('[{"hash": "' || iss.target_hash || '"}]')::jsonb AND created_at > current_timestamp - interval '14d')
-      ELSE (0, NOW()::TEXT)
+      ELSE (0::BIGINT, NOW()::TEXT)
     END as req_count
     FROM apis.issues iss WHERE project_id = ?  and created_at > current_timestamp - interval $report_interval
     ORDER BY req_count desc limit 20
